chore(client): [CLI-024] :truck: Move InfoInForm in form package
