+++ client/core/src/main/java/fr/esgi/pa/alliodesk/core/form/InfoInForm.java
@@ -1,4 +1,4 @@
-package fr.esgi.pa.alliodesk.core;
+package fr.esgi.pa.alliodesk.core.form;
 
 public class InfoInForm {
     private String name,
