+++ client/core/src/main/java/fr/esgi/pa/alliodesk/core/request/AllioErrorRequest.java
@@ -1,6 +1,6 @@
 package fr.esgi.pa.alliodesk.core.request;
 
-import fr.esgi.pa.alliodesk.core.InfoInForm;
+import fr.esgi.pa.alliodesk.core.form.InfoInForm;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpPost;
 
@@ -8,10 +8,12 @@ import java.io.IOException;
 
 public class AllioErrorRequest extends ApiRequest {
     private final InfoInForm errorForm;
-    public AllioErrorRequest( String message) {
+
+    public AllioErrorRequest(String message) {
         this.errorForm = InfoInForm.build()
                 .withMessage(message);
     }
+
     @Override
     public int requestToServe() {
         try {
